+++ src/components/CreateUser.js
@@ -1,6 +1,27 @@
-import React from "react";
+import React, { useRef, useContext } from "react";
+import useInputs from "hooks/useInputs";
+import { UserDispatch } from "App";
 
-const CreateUser = ({ username, email, onChange, onCreate }) => {
+const CreateUser = () => {
+  const [{ username, email }, onChange, reset] = useInputs({
+    username: "",
+    email: "",
+  });
+  const nextId = useRef(4);
+  const dispatch = useContext(UserDispatch);
+
+  const onCreate = () => {
+    dispatch({
+      type: "CREATE_USER",
+      user: {
+        id: nextId.current,
+        username,
+        email,
+      },
+    });
+    reset();
+    nextId.current += 1;
+  };
   return (
     <div>
       <input
