+++ src/components/UserList.js
@@ -1,36 +1,30 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useContext } from "react";
+import { UserDispatch } from "App";
 
-const User = React.memo(({ user, onRemove, onToggle }) => {
-  useEffect(() => {
-    console.log(user);
-  }, [user]);
+const User = React.memo(({ user }) => {
+  const dispatch = useContext(UserDispatch);
 
   return (
     <div>
       <b
         style={{ cursor: "pointer", color: user.active ? "green" : "black" }}
-        onClick={() => onToggle(user.id)}
+        onClick={() => dispatch({ type: "TOGGLE_USER", id: user.id })}
       >
         {user.username}
       </b>
       <span>({user.email})</span>
-      <button onClick={() => onRemove(user.id)}>삭제</button>
+      <button onClick={() => dispatch({ type: "REMOVE_USER", id: user.id })}>
+        삭제
+      </button>
     </div>
   );
 });
 
-const UserList = ({ users, onRemove, onToggle }) => {
+const UserList = ({ users }) => {
   return (
     <div>
       {users.map((user) => {
-        return (
-          <User
-            user={user}
-            key={user.id}
-            onRemove={onRemove}
-            onToggle={onToggle}
-          ></User>
-        );
+        return <User user={user} key={user.id}></User>;
       })}
     </div>
   );
